+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Fixes:
 Features:
 
 * Add option to set default-value for execution-prompt choices/dropdown [#93](https://github.com/O-X-L/ansible-webui/issues/93)
+* Add buttons to log-view for job-rerun and job-edit [#94](https://github.com/O-X-L/ansible-webui/issues/94)
 
 ----